requirements.txt
# wrapper - minimal dependencies
# Only PyYAML is required (standard library handles the rest)

PyYAML>=6.0
